change to f32s

+++ src/camera.rs
@@ -5,16 +5,16 @@ use vector::Vector;
 use point::Point;
 
 pub enum Camera {
-    OrthoCamera(Transform, Box<Film>, f64),
-    PerspectiveCamera(Transform, Box<Film>, f64),
+    OrthoCamera(Transform, Box<Film>, f32),
+    PerspectiveCamera(Transform, Box<Film>, f32),
 }
 
 impl Camera {
-    pub fn new_ortho(f : Box<Film>, scale : f64) -> Camera {
+    pub fn new_ortho(f : Box<Film>, scale : f32) -> Camera {
         OrthoCamera(Transform::identity(), f, scale)
     }
 
-    pub fn new_perspective(f : Box<Film>, fov_y : f64) -> Camera {
+    pub fn new_perspective(f : Box<Film>, fov_y : f32) -> Camera {
         PerspectiveCamera(Transform::identity(), f, fov_y)
     }
 
@@ -25,32 +25,32 @@ impl Camera {
         }
     }
 
-    pub fn cast(&self, fx : f64, fy : f64) -> Ray {
+    pub fn cast(&self, fx : f32, fy : f32) -> Ray {
         match self {
             &OrthoCamera(ref t, ref f, s) => {
-                let fw = f.width as f64;
-                let fh = f.height as f64;
-                let x = fx - (fw / 2f64);
-                let y = fy - (fh / 2f64);
+                let fw = f.width as f32;
+                let fh = f.height as f32;
+                let x = fx - (fw / 2f32);
+                let y = fy - (fh / 2f32);
                 let d = Vector::unit_z();
-                let o = Point::new(x*s, y*s, 0f64);
+                let o = Point::new(x*s, y*s, 0f32);
                 Ray::new(&o, &d).transform(&t.inverse())
             },
             &PerspectiveCamera(ref t, ref f, fov_y) => {
-                let fw = f.width as f64;
-                let fh = f.height as f64;
-                let x = (fx / fw) * 2f64 - 1f64;
-                let y = (fy / fh) * 2f64 - 1f64;
-                let sx = (fov_y / 2f64).tan() * (fw / fh);
-                let sy = (fov_y / 2f64).tan();
-                let d = Vector::new(x * sx, y * sy, 1f64).normalize();
+                let fw = f.width as f32;
+                let fh = f.height as f32;
+                let x = (fx / fw) * 2f32 - 1f32;
+                let y = (fy / fh) * 2f32 - 1f32;
+                let sx = (fov_y / 2f32).tan() * (fw / fh);
+                let sy = (fov_y / 2f32).tan();
+                let d = Vector::new(x * sx, y * sy, 1f32).normalize();
                 let o = Point::origin();
                 Ray::new(&o, &d).transform(&t.inverse())
             },
         }
     }
 
-    pub fn receive(&mut self, fx : f64, fy : f64, p : u8) {
+    pub fn receive(&mut self, fx : f32, fy : f32, p : u8) {
         match self {
             &OrthoCamera(_, ref mut f, _) => f.add_sample(fx, fy, p),
             &PerspectiveCamera(_, ref mut f, _) => f.add_sample(fx, fy, p),
